feat(threat-hunter): integrate self-contained AgentEngineSandboxCodeExecutor and deploy to Argolis
- Move code executor factory into self-contained agent_a2a_threat_hunter module
- Implement __reduce__ and __deepcopy__ lock handling for cloudpickle serialization
- Default get_code_executor('auto') to AgentEngineSandboxCodeExecutor to prevent Gemini built-in tool collisions with custom MCP tools
- Successfully deployed Threat Hunter Reasoning Engine to Argolis (projects/813924125873/locations/us-central1/reasoningEngines/4573376284536602624)
- Verified live end-to-end query execution with session 2660560887856758784
