+++ .gitignore
@@ -9,6 +9,7 @@ __pycache__/
 implementation_plan.md
 .env*
 .env.secops.adc
+.env.argolis
 investigations/*
 !investigations/similarity_reports/
 *.json
